Rails: move fever_api under commands
This makes `commands/` a more consistent place to put callable objects
and will work a little more nicely with Zeitwerk.
