+++ app.rb
@@ -17,12 +17,20 @@ require "sprockets-helpers"
 require "securerandom"
 
 require_relative "app/commands/feeds/import_from_opml"
+require_relative "app/commands/fever_api/response"
 require_relative "app/commands/stories/mark_all_as_read"
+require_relative "app/commands/stories/mark_as_read"
+require_relative "app/commands/stories/mark_as_starred"
+require_relative "app/commands/stories/mark_as_unread"
+require_relative "app/commands/stories/mark_as_unstarred"
+require_relative "app/commands/stories/mark_feed_as_read"
+require_relative "app/commands/stories/mark_group_as_read"
 require_relative "app/commands/users/create_user"
 require_relative "app/commands/users/sign_in_user"
-require_relative "app/fever_api/response"
 require_relative "app/repositories/story_repository"
 require_relative "app/repositories/user_repository"
+require_relative "app/repositories/feed_repository"
+require_relative "app/repositories/group_repository"
 require_relative "app/tasks/fetch_feeds"
 require_relative "config/asset_pipeline"
