+++ app/commands/fever_api/authentication.rb
