+++ app/commands/fever_api/read_favicons.rb
