+++ app/commands/fever_api/read_feeds_groups.rb
@@ -1,7 +1,5 @@
 # frozen_string_literal: true
 
-require_relative "../repositories/feed_repository"
-
 module FeverAPI
   class ReadFeedsGroups
     def initialize(options = {})
