+++ app/commands/fever_api/read_groups.rb
@@ -1,7 +1,5 @@
 # frozen_string_literal: true
 
-require_relative "../repositories/group_repository"
-
 module FeverAPI
   class ReadGroups
     def initialize(options = {})
