+++ app/commands/fever_api/read_items.rb
@@ -1,7 +1,5 @@
 # frozen_string_literal: true
 
-require_relative "../repositories/story_repository"
-
 module FeverAPI
   class ReadItems
     def initialize(options = {})
