+++ app/commands/fever_api/write_mark_feed.rb
@@ -1,7 +1,5 @@
 # frozen_string_literal: true
 
-require_relative "../commands/stories/mark_feed_as_read"
-
 module FeverAPI
   class WriteMarkFeed
     def initialize(options = {})
