+++ app/commands/fever_api/write_mark_item.rb
@@ -1,10 +1,5 @@
 # frozen_string_literal: true
 
-require_relative "../commands/stories/mark_as_read"
-require_relative "../commands/stories/mark_as_unread"
-require_relative "../commands/stories/mark_as_starred"
-require_relative "../commands/stories/mark_as_unstarred"
-
 module FeverAPI
   class WriteMarkItem
     def initialize(options = {})
